Factor out generic socket scheme.
